Move the data iter tests.
